+++ bioblend/_tests/TestGalaxyWorkflows.py
@@ -72,7 +72,6 @@ class TestGalaxyWorkflows(GalaxyTestBase.GalaxyTestBase):
         invocation = self.gi.workflows.show_invocation(workflow_id, invocation_id)
         assert invocation["state"] == "scheduled"
 
-    @test_util.skip_unless_galaxy("release_19.01")
     def test_invoke_workflow_parameters_normalized(self):
         path = test_util.get_abspath(os.path.join("data", "paste_columns_subworkflow.ga"))
         workflow_id = self.gi.workflows.import_workflow_from_local_path(path)["id"]
@@ -194,7 +193,6 @@ class TestGalaxyWorkflows(GalaxyTestBase.GalaxyTestBase):
         assert len(wf_data["steps"]) == 3
         assert wf_data["inputs"] is not None
 
-    @test_util.skip_unless_galaxy("release_18.05")
     def test_update_workflow_name(self):
         path = test_util.get_abspath(os.path.join("data", "paste_columns.ga"))
         wf = self.gi.workflows.import_workflow_from_local_path(path)
@@ -268,7 +266,6 @@ class TestGalaxyWorkflows(GalaxyTestBase.GalaxyTestBase):
             assert wf1["steps"][str(i)]["type"] == wf2["steps"][str(i)]["type"]
             assert wf1["steps"][str(i)]["tool_id"] == wf2["steps"][str(i)]["tool_id"]
 
-    @test_util.skip_unless_galaxy("release_18.09")
     def test_show_versions(self):
         path = test_util.get_abspath(os.path.join("data", "paste_columns.ga"))
         wf = self.gi.workflows.import_workflow_from_local_path(path)